requirements.txt
flask
gunicorn
torch
torchvision
transformers
scikit-learn
numpy
pillow
opencv-python-headless
requests